refactor(api): move websocket host adapter to lib/server
Keep index.ts to route registration: the Vercel-vs-Bun adapter
selection, the WebSocketServer/serve() wiring, and the dual
server-export shape move into @/lib/server (upgradeWebSocket +
createServer). No behavior change; re-smoked both branches (Bun.serve
and VERCEL=1 node+ws) streaming snapshot + heartbeat, and REST still
200s on the node export. Docs and the api-endpoint skill point at the
new module.
