+++ api/hono/src/index.ts
@@ -1,29 +1,20 @@
-import { serve, upgradeWebSocket as nodeUpgradeWebSocket } from "@hono/node-server"
 import { site } from "@packages/config/site"
 import { getBuildVersion } from "@packages/env"
 import { env } from "@packages/env/api-hono"
 import { Scalar } from "@scalar/hono-api-reference"
 import { Hono } from "hono"
 import { describeRoute, openAPIRouteHandler, resolver } from "hono-openapi"
-import { upgradeWebSocket as bunUpgradeWebSocket, websocket } from "hono/bun"
 import { cors } from "hono/cors"
 import { HTTPException } from "hono/http-exception"
 import { logger } from "hono/logger"
-import { WebSocketServer } from "ws"
 import { z } from "zod"
 
 import { errorHandler, globalErrorResponses, jsonError } from "@/lib/error"
+import { createServer, upgradeWebSocket } from "@/lib/server"
 import { rateLimiterMiddleware } from "@/middlewares"
 import { agentsRouter, authRouter, v1Router, waitlistRouter } from "@/routers"
 
 const BUILD_VERSION = getBuildVersion()
-
-// Vercel Functions can't run Bun.serve(), so on Vercel we serve WebSockets through the Node adapter (@hono/node-server + ws); everywhere else (local, Docker/self-host) Bun.serve() owns the socket via hono/bun.
-const onVercel = Boolean(process.env.VERCEL)
-// Both adapters accept the same handler factory; the cast collapses their otherwise non-unionable signatures to one callable type.
-const upgradeWebSocket = onVercel
-  ? (nodeUpgradeWebSocket as typeof bunUpgradeWebSocket)
-  : bunUpgradeWebSocket
 
 const app = new Hono()
 
@@ -182,15 +173,5 @@ socket.addEventListener("message", (event) => {
 export type AppType = typeof routes
 export type { ErrorCode } from "@/lib/error"
 
-// On Vercel, export the Node http.Server so the platform drives all traffic including the WebSocket upgrade; it binds PORT when Vercel sets it, else HONO_PORT (e.g. forced locally). Elsewhere, export the Bun.serve() shape so Bun owns fetch + the socket.
-export default onVercel
-  ? serve({
-      fetch: app.fetch,
-      port: Number(process.env.PORT) || env.HONO_PORT,
-      websocket: { server: new WebSocketServer({ noServer: true }) },
-    })
-  : {
-      port: env.HONO_PORT,
-      fetch: app.fetch,
-      websocket,
-    }
+// Bun.serve() shape locally and self-hosted, Node http.Server on Vercel; see @/lib/server.
+export default createServer(app)
