+++ web/next/content/docs/getting-started/type-safe-api.mdx
@@ -77,18 +77,33 @@ Every route decorated with `describeRoute` shows up in an interactive [Scalar](h
 
 Hono serves [WebSockets](https://hono.dev/docs/helpers/websocket) and the same typed client reaches them. The health endpoint ships a live example at `/api/health/ws` that pushes a status snapshot on connect and a heartbeat every few seconds.
 
-The route is a `GET` upgraded with `upgradeWebSocket`, but the socket owner differs by host. Bun (local, Docker/self-host) runs `Bun.serve()`, so the socket comes from `hono/bun`. Vercel Functions can't run `Bun.serve()`, so on Vercel the upgrade goes through the Node adapter (`@hono/node-server` + `ws`), which Vercel [drives natively](https://vercel.com/docs/functions/websockets). The starter picks the adapter at boot from `process.env.VERCEL` and exports the matching server:
+The route is a `GET` upgraded with `upgradeWebSocket`, but the socket owner differs by host. Bun (local, Docker/self-host) runs `Bun.serve()`, so the socket comes from `hono/bun`. Vercel Functions can't run `Bun.serve()`, so on Vercel the upgrade goes through the Node adapter (`@hono/node-server` + `ws`), which Vercel [drives natively](https://vercel.com/docs/functions/websockets). `lib/server.ts` picks the adapter at boot from `process.env.VERCEL` and builds the matching server export, keeping `index.ts` to route registration:
 
 ```ts
-// api/hono/src/index.ts
+// api/hono/src/lib/server.ts
 import { serve, upgradeWebSocket as nodeUpgradeWebSocket } from "@hono/node-server"
 import { upgradeWebSocket as bunUpgradeWebSocket, websocket } from "hono/bun"
 import { WebSocketServer } from "ws"
 
 const onVercel = Boolean(process.env.VERCEL)
-const upgradeWebSocket = onVercel
+
+export const upgradeWebSocket = onVercel
   ? (nodeUpgradeWebSocket as typeof bunUpgradeWebSocket)
   : bunUpgradeWebSocket
+
+export const createServer = (app: Hono) =>
+  onVercel
+    ? serve({
+        fetch: app.fetch,
+        port: Number(process.env.PORT) || env.HONO_PORT,
+        websocket: { server: new WebSocketServer({ noServer: true }) },
+      })
+    : { port: env.HONO_PORT, fetch: app.fetch, websocket }
+```
+
+```ts
+// api/hono/src/index.ts
+import { createServer, upgradeWebSocket } from "@/lib/server"
 
 const routes = app.basePath("/api").get(
   "/health/ws",
@@ -99,13 +114,7 @@ const routes = app.basePath("/api").get(
   })),
 )
 
-export default onVercel
-  ? serve({
-      fetch: app.fetch,
-      port: Number(process.env.PORT) || env.HONO_PORT,
-      websocket: { server: new WebSocketServer({ noServer: true }) },
-    })
-  : { port: env.HONO_PORT, fetch: app.fetch, websocket }
+export default createServer(app)
 ```
 
 On the client, the RPC object exposes `$ws()` on the route, so the connection URL stays derived from the same configured API base (`http` becomes `ws`):
